requirements.txt
streamlit==1.35.0
tavily-python==0.3.3
python-dotenv==1.0.1
groq==0.9.0
httpx==0.27.0
